+++ src/types/execution.ts
@@ -90,4 +90,10 @@ export type Execution = {
   workflowDefinition: WorkflowDef;
   reasonForIncompletion?: string;
   workflowEngine?: string;
+  workflowVersion?: string;
+  startTime?: number;
+  endTime?: number;
+  parentWorkflowTaskId?: string;
+  externalInputPayloadStoragePath?: string; // TODO: can be removed for conductor 4
+  externalOutputPayloadStoragePath?: string; // TODO: can be removed for conductor 4
 };
